Implemented all necessary components for db3, including queries. Tested implementation, queries work fine. Some refactoring, deleting old classes. Update postman collection.

+++ backend/src/main/java/com/Project/backend/model/heartDiseasePredict.java
@@ -20,9 +20,8 @@ public class heartDiseasePredict {
     private Integer totChol;
     private Integer sysBP;
     private Integer diaBP;
-    private Integer BMI;
+    private Double BMI;
     private Integer heartRate;
     private Integer glucose;
     private Integer TenYearCHD;
-
 }
